Split catalog db_document.rs into marshaling submodules
db_document/ now holds video_out (row/manifest/catalog-entry to API
DTO conversion), row_manifest (per-row manifest document assembly),
and build (public catalog document construction). Pure move with
module-root glob re-exports.
